Auto-generated SDK

+++ waf/src/models/waf_events200.rs
@@ -11,12 +11,12 @@
 
 
 
-#[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct WafEvents200 {
     #[serde(rename = "results", skip_serializing_if = "Option::is_none")]
     pub results: Option<Vec<serde_json::Value>>,
     #[serde(rename = "schema_version", skip_serializing_if = "Option::is_none")]
-    pub schema_version: Option<i32>,
+    pub schema_version: Option<i64>,
 }
 
 impl WafEvents200 {
